Consolidate matching declarations
