--- feature/apps/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/apps/impl/components/PermissionRationaleBottomSheet.kt
@@ -1,82 +1,0 @@
-package sk.styk.martin.apkanalyzer.feature.apps.impl.components
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import sk.styk.martin.apkanalyzer.core.uilibrary.components.BottomSheet
-import sk.styk.martin.apkanalyzer.core.uilibrary.components.Button
-import sk.styk.martin.apkanalyzer.core.uilibrary.components.Icon
-import sk.styk.martin.apkanalyzer.core.uilibrary.components.Text
-import sk.styk.martin.apkanalyzer.core.uilibrary.icons.ApkAnalyzerIcons
-import sk.styk.martin.apkanalyzer.core.uilibrary.theme.ApkAnalyzerTheme
-import sk.styk.martin.apkanalyzer.core.uilibrary.theme.AppTheme
-
-@Composable
-internal fun PermissionRationaleBottomSheet(
-    title: String,
-    description: String,
-    openSettingsLabel: String,
-    onOpenSettings: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    BottomSheet(onDismiss = onDismiss) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 24.dp)
-                .padding(bottom = 24.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Row(
-                verticalAlignment = Alignment.Top,
-                horizontalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                Icon(
-                    imageVector = ApkAnalyzerIcons.Lock,
-                    tint = AppTheme.colors.onSurfaceVariant,
-                    modifier = Modifier.size(20.dp),
-                )
-                Text(
-                    text = title,
-                    style = AppTheme.typography.titleMedium,
-                    color = AppTheme.colors.onSurface,
-                )
-            }
-            Text(
-                text = description,
-                style = AppTheme.typography.bodyMedium,
-                color = AppTheme.colors.onSurfaceVariant,
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            Button(
-                text = openSettingsLabel,
-                onClick = onOpenSettings,
-                modifier = Modifier.fillMaxWidth(),
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun PermissionRationaleBottomSheetPreview() {
-    ApkAnalyzerTheme {
-        PermissionRationaleBottomSheet(
-            title = "Allow usage access",
-            description = "These filters check when you last opened each app. Grant usage access in Settings so the app can identify recently active and idle apps.",
-            openSettingsLabel = "Open settings",
-            onOpenSettings = {},
-            onDismiss = {},
-        )
-    }
-}
